Add more classes to make the code more OOP
